chore: remove hotfix for Postgres
We are using the letsql Postgres backend,
there is no need to hotfix the ibis one.
